Challenges check teacher and student

+++ app/Http/Controllers/GroupsController.php
@@ -16,13 +16,18 @@ class GroupsController extends Controller
     public function index($code){
        
         $class = Classroom::where('code', '=', $code)->with('grouping.groups', 'students.groups')->firstorFail();
+        $this->authorize('view', $class);
         return view('groups.index', compact('class'));
         
     }
 
     public function update($code) {
+        $class = Classroom::where('code', '=', $code)->firstorFail();
+        $this->authorize('update', $class);
         foreach (request()->dropGroups as $group) {
             $groupObj = Group::findOrFail($group['id']);
+            if($groupObj->grouping->classroom_id != $class->id)
+                continue;
             $groupObj->students()->detach($groupObj->students);
             foreach ($group['children'] as $std) {
                 $groupObj->students()->attach($std['id']);
@@ -32,6 +37,7 @@ class GroupsController extends Controller
     
     public function store($code) {
         $class = Classroom::where('code', '=', $code)->firstorFail();
+        $this->authorize('update', $class);
         $data = request()->validate([
             'groupsNumber' => ['numeric'],
         ]);
