+++ app/Http/Controllers/InventoryController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Classroom;
 use App\Student;
 
 class InventoryController extends Controller
@@ -12,10 +13,11 @@ class InventoryController extends Controller
         $this->middleware('verified');
     }
     
-    public function update($code) {
-        // $class = Classroom::where('code', '=', $code)->firstOrFail();
+    public function update() {
         $request = request();
         $student = Student::findOrFail($request->id);
+        $class = Classroom::where('id', '=', $student->classroom->classroom_id)->firstOrFail();
+        $this->authorize('update', $class);
         if($request->count <= 0) {
             $student->items()->detach($request->item_id);
         } else {
